fix(taxis): wire or remove three dead config surfaces
timeouts.llmCallSecs: deleted. Zero production readers existed; real LLM
call timeouts run through providerBehavior.nonStreamingTimeoutSecs
(aletheia/src/runtime/setup.rs). Deleted the field (timeouts.rs), its
koina::defaults::TIMEOUT_SECONDS default source (now itself unread
anywhere), its registry.rs ParameterSpec entry, its validate.rs range
check, its CONFIGURATION.md row, its instance.example line, and the
config_tests/deployment_wave1.rs assertions that pinned it. Added a doc
pointer from the timeouts section to the surviving timeout knob.

[tuning]: wired. nous::pipeline::run_session_tuning_proposer built a
fresh TuningConfig{enabled: true, ..default()} every call, discarding
every operator-set field including the documented global kill switch.
Added PipelineConfig.tuning (resolved once from AletheiaConfig.tuning
in aletheia::runtime::nous_config::build_nous_runtime_config), threaded
it through run_pipeline's existing pipeline_config parameter, and
construct TuningProposer from that instead. Ephemeral sub-agent spawns
(spawn_svc.rs) get TuningConfig::default() rather than inheriting the
top-level config, consistent with their existing no-training-capture
posture. Added a regression test proving all four TuningConfig fields
now cascade from AletheiaConfig into PipelineConfig.

BACKOFF_FACTOR: wired. hermeneus::retry::RetryPolicy::delay hardcoded
factor: 2 in its ExponentialJitter strategy while BACKOFF_FACTOR sat
unread outside its own pinning test. Changed the constant's type to u32
(matching BackoffStrategy::ExponentialJitter::factor) and pass it
directly instead of the literal. Fixed the one downstream consumer
(hermeneus/tests/public_api.rs) that multiplied it against a u64.

Investigated but NOT changed: nous_behavior.self_audit_event_threshold
(cited in the issue as "same class" evidence). SelfAuditor::new() does
read NousBehaviorConfig::default() for this field, but
run_post_turn_self_audit calls SelfAuditor::run_audit directly with a
hardcoded AuditTrigger::EventBased{after_n_actions: 1} -- it never
calls SelfAuditor::record_action, the method that actually consults
event_threshold. The field is inert in production regardless of its
value; threading the resolved config into it would not change runtime
behavior and would misrepresent the fix. Left for separate scoping
since the real defect (record_action is never invoked) is a behavior
change beyond a config-wiring fix.

Also found but out of scope for this issue's title: nous_behavior's
gc_interval_secs and max_spawned_tasks show the identical pattern
(gc_interval_secs read via NousBehaviorConfig::default() in
tasks/gc.rs; max_spawned_tasks has zero readers at all).

+++ crates/hermeneus/tests/public_api.rs
@@ -52,7 +52,7 @@ mod model_constants {
                 "exponential backoff requires factor >= 2"
             );
             assert!(
-                BACKOFF_MAX_MS >= BACKOFF_BASE_MS * BACKOFF_FACTOR,
+                BACKOFF_MAX_MS >= BACKOFF_BASE_MS * u64::from(BACKOFF_FACTOR),
                 "max must allow at least one full backoff step"
             );
         }
